docs first pass and examples
