+++ docs/content/docs/examples/visgym.mdx
@@ -1,0 +1,159 @@
+---
+title: "Multi-Turn Multi-Image VLM RL on VisGym"
+---
+
+<Callout type="info">
+Uses the `fsdp` backend. See the [Vision-Language RL tutorial](../tutorials/vision_language_rl) for the VLM setup background (required flags, vllm nightly, support matrix).
+</Callout>
+
+In this example we train a vision-language model to solve VisGym tasks — gymnasium-style environments (e.g. `maze_2d/easy`) where every environment step returns a **new image observation**. The model sees an image, reasons, emits an action, and receives the next rendered image as part of the next user turn. This is the multi-image multi-turn VLM pattern described in the tutorial.
+
+Source code: `examples/train/visgym/`.
+
+## Task Overview
+
+VisGym environments are `gymnasium.Env` instances that return rendered images as observations. The example defaults to `maze_2d/easy`, a 2D navigation maze where:
+
+- The env renders the maze (walls, agent position, goal) as an RGB image.
+- The model chooses a direction to move.
+- The env steps, re-renders, and sends back the new state.
+- The episode ends when the agent reaches the goal, hits a termination condition, or runs out of turns.
+
+Reward is sparse — delivered at the terminal step based on whether the task was solved.
+
+## Two Recipes
+
+VisGym ships with two training recipes that differ in how the model is asked to output actions:
+
+### `run_visgym_from_sft.sh` — tuple-action, SFT warm-start
+
+Starts from a pre-SFT'd Qwen3-VL checkpoint that already emits structured output:
+
+```
+<observation>I see a maze with the agent at (1,1) and the goal at (3,3)...</observation>
+<justification>The goal is east-southeast, so I should move east first.</justification>
+<action>('move', 0)</action>
+```
+
+Reward combines task success with a **format reward** (0.8 task + 0.2 format) so the model preserves the SFT output structure during RL. KL regularization is off.
+
+See `examples/train/visgym/env_sft.py` for the action-parsing logic (a balanced-parenthesis walker with several fallback strategies — SFT models sometimes emit markdown-wrapped or JSON-shaped actions).
+
+### `run_visgym_from_instruct.sh` — keyword-action, from vanilla instruct
+
+Starts from `Qwen/Qwen3-VL-8B-Instruct` without an SFT warm-start. The action format is a simple keyword tag:
+
+```
+<action>left</action>
+```
+
+where the keyword is one of `left`, `right`, `up`, `down`, `stop`. Reward is **task-only** (no format shaping) and KL regularization is on (`kl_loss_coef=0.005`) to keep the policy anchored to the instruct distribution.
+
+See `examples/train/visgym/env_instruct.py` — the action extractor is a simple regex match for `<action>keyword</action>` tags. The env sets `relaxed=True` on the underlying VisGym env so the gymnasium env accepts keyword inputs.
+
+## Env Contract
+
+Both recipes share the multi-image-per-turn contract. On each `step()`:
+
+```python
+obs, reward, terminated, truncated, info = self.visgym_env.step(extracted)
+if not done:
+    image = self.visgym_env.render()               # new RGB frame
+    obs_msg = make_image_message(feedback, image)   # {role, content=[text, image_url]}
+    observations = [obs_msg]
+```
+
+The observation is an OpenAI-format message with both a text feedback string and a base64-encoded image in an `image_url` content part. These accumulate in the conversation, so by turn N the model has seen N+1 images.
+
+`examples/train/visgym/utils.py` provides `make_image_message()`, which handles the RGB-array → data-URI conversion consistently across both recipes.
+
+## Data Preparation
+
+Because VisGym generates its prompts (and images) dynamically from the env's state, the dataset is just a list of "tickets" that tell SkyRL which env to run. It's auto-generated on first run:
+
+```bash
+uv run examples/train/visgym/dataset.py \
+  --env_id maze_2d/easy \
+  --num_rows 256 \
+  --output_dir ~/data/visgym_maze_2d_easy
+```
+
+Each row carries `env_class=visgym`, `visgym_env_id`, and optionally a deterministic `seed` for evaluation.
+
+## Training Configuration
+
+Key overrides from both scripts:
+
+```bash title="examples/train/visgym/run_visgym_from_sft.sh (excerpt)"
+_SKYRL_USE_NEW_INFERENCE=1 uv run --isolated --extra fsdp \
+  python examples/train/visgym/entrypoint_sft.py \
+  trainer.policy.model.path="$MODEL_PATH" \         # SFT checkpoint
+  trainer.strategy=fsdp2 \
+  # VLM flags
+  generator.is_vlm=true \
+  generator.batched=false \
+  trainer.use_sample_packing=false \
+  # Multi-turn rollouts
+  generator.max_turns=15 \
+  generator.inference_engine.async_engine=true \
+  generator.sampling_params.temperature=0.7 \
+  # Algorithm — task+format reward, no KL
+  trainer.algorithm.advantage_estimator="grpo" \
+  trainer.algorithm.use_kl_loss=false \
+  environment.env_class=visgym \
+```
+
+```bash title="examples/train/visgym/run_visgym_from_instruct.sh (excerpt)"
+_SKYRL_USE_NEW_INFERENCE=1 uv run --isolated --extra fsdp \
+  python examples/train/visgym/entrypoint_instruct.py \
+  trainer.policy.model.path="Qwen/Qwen3-VL-8B-Instruct" \
+  trainer.strategy=fsdp2 \
+  # Same VLM flags as above
+  generator.is_vlm=true \
+  generator.batched=false \
+  trainer.use_sample_packing=false \
+  # More turns, higher temperature, KL regularization on
+  generator.max_turns=18 \
+  generator.sampling_params.temperature=1 \
+  trainer.algorithm.use_kl_loss=true \
+  trainer.algorithm.kl_loss_coef=0.005 \
+  environment.env_class=visgym \
+```
+
+Notable differences: the instruct recipe runs more turns (18 vs 15), samples hotter (temperature 1 vs 0.7), and uses KL to stay close to the base distribution — all compensations for not having an SFT warm-start.
+
+## Launching
+
+```bash
+# Instruct recipe — no SFT checkpoint required
+bash examples/train/visgym/run_visgym_from_instruct.sh
+
+# SFT recipe — point MODEL_PATH at your checkpoint
+MODEL_PATH=/path/to/your/sft_ckpt bash examples/train/visgym/run_visgym_from_sft.sh
+```
+
+Tested on 1× 8× H100 node.
+
+## Example Rollout
+
+![VisGym rollout](/images/examples/visgym_rollout.gif)
+
+<Callout type="info">
+Placeholder — a real multi-image multi-turn rollout (per-turn maze frame updates as the agent navigates) will replace this asset.
+</Callout>
+
+## Results
+
+![VisGym from-SFT reward curve](/images/examples/visgym_sft_reward.png)
+
+![VisGym from-instruct reward curve](/images/examples/visgym_instruct_reward.png)
+
+<Callout type="info">
+Placeholders — real reward curves from each recipe will replace these assets.
+</Callout>
+
+## What's Next
+
+- **[Geometry-3K example](geometry3k)** — single-image multi-turn VLM RL with tool calling.
+- **[Vision-Language RL tutorial](../tutorials/vision_language_rl)** — shared setup background.
+- **[Creating a New Environment](../tutorials/new_env)** — build your own env.
