+++ docs/content/docs/tutorials/vision_language_rl.mdx
@@ -1,0 +1,84 @@
+---
+title: "Vision-Language RL in SkyRL"
+---
+
+
+SkyRL supports multi-turn RL training on vision-language models (VLMs) — the policy reads image observations, reasons, and emits actions or tool calls, with images flowing through the conversation across turns. This guide covers the setup common to all VLM training runs in SkyRL: enabling the VLM code path, the required flags, the dataset record shape, and the current support matrix. For concrete runnable recipes, see the [Geometry-3K](../examples/geometry3k) and [VisGym](../examples/visgym) example pages.
+
+## Requirements
+
+<Callout type="warn">
+**TODO: vllm nightly install command.** VLM training requires a nightly build of vLLM that includes the multimodal rendering support used by SkyRL's new inference stack. The repo pins `vllm==0.19.0` in `pyproject.toml` for non-VLM workflows, but VLM runs need a newer wheel. Replace this callout with the exact install command before shipping.
+</Callout>
+
+**Hardware.** The examples in this guide were tested on a single node with 8× H100 GPUs. VLM runs are memory-heavy (image tokens expand dramatically during encoding) — calibrate `gpu_memory_utilization` and `max_model_len` if you're working with a different hardware profile.
+
+## Turning on VLM mode
+
+Two flags are what actually switch a training run into VLM mode:
+
+```bash
+generator.is_vlm=true \
+trainer.use_sample_packing=false \
+```
+
+- **`generator.is_vlm=true`** — the one real VLM switch. Wires in the VLM-aware generator so that image content parts in the prompt are routed through the vision path of the inference engine.
+- **`trainer.use_sample_packing=false`** — sample packing is currently unsupported for VLM training (see callout below).
+
+<Callout type="info">
+**Sample packing is deferred for VLM.** Each VLM family (Qwen-VL, InternVL, etc.) processes images and positional embeddings differently, and packing across heterogeneous image counts per sample would need per-model handling. Support is expected in a future release; until then, set `trainer.use_sample_packing=false` for any VLM run.
+</Callout>
+
+You'll also want `generator.batched=false` for multi-turn rollouts, but this isn't VLM-specific — the same flag appears in non-VLM multi-turn recipes like [Text-to-SQL](../examples/multi_turn_text2sql).
+
+## Dataset record shape
+
+VLM training datasets embed images directly in the prompt as `image_url` content parts, using base64 data URIs:
+
+```python
+{
+    "prompt": [
+        {
+            "role": "user",
+            "content": [
+                {"type": "image_url", "image_url": {"url": "data:image/jpeg;base64,<...>"}},
+                {"type": "text", "text": "What is shown in this image?"},
+            ],
+        },
+    ],
+    "env_class": "geometry3k",
+    "reward_spec": {"method": "rule", "ground_truth": "..."},
+}
+```
+
+A prompt can have multiple `image_url` parts. See `examples/train/geometry3k/geometry_3k_dataset.py` for a full dataset preparation script that converts PIL images to data URIs.
+
+For environments that generate images dynamically at each turn (like VisGym), the dataset contains only stub prompts — the env builds the actual multimodal message in `init()` and `step()`. See `examples/train/visgym/dataset.py`.
+
+## Interaction patterns
+
+SkyRL handles two distinct VLM interaction patterns:
+
+**Text-only observations.** The initial prompt includes image(s), but subsequent environment observations are text (e.g. tool-call feedback). The [Geometry-3K example](../examples/geometry3k) uses this pattern — the model sees a geometry diagram once, then iterates with text feedback from a `calc_score` tool.
+
+**Image-returning observations.** The environment returns a new image observation at every step. The [VisGym example](../examples/visgym) uses this pattern — after each model action, `env.step()` renders a new image of the updated world state, which is passed back as the next user turn's content.
+
+The generator handles both patterns transparently once `generator.is_vlm=true` is set. The environment contract determines which pattern you're in: return observations with `image_url` content parts for image-returning envs, or plain text strings for text-only feedback.
+
+## Support matrix
+
+| Feature | Status | Notes |
+|---|---|---|
+| FSDP backend | Supported | Used by both example recipes. |
+| Megatron backend | Not supported | VLM path has not been wired through Megatron yet. |
+| Sample packing | Deferred | Per-model image/positional-embedding handling required. |
+| Batched generation | Not supported | Use `generator.batched=false` + async engine for multi-turn. |
+| Qwen3-VL | Tested | `Qwen/Qwen3-VL-8B-Instruct` used in both examples. |
+| Multi-image per turn | Supported | See VisGym example. |
+| Hardware tested | 1× 8× H100 | |
+
+## What's next
+
+- **[Geometry-3K example](../examples/geometry3k)** — single-image multi-turn with tool calling.
+- **[VisGym example](../examples/visgym)** — multi-image multi-turn, with a new image per step.
+- **[Creating a New Environment](new_env)** — build your own env on top of the VLM flags in this guide.
