+++ examples/train/visgym/run_visgym_from_sft.sh
@@ -2,26 +2,27 @@
 set -euo pipefail
 set -x
 
-# Full RL training run: VLM RL on VisGym maze_2d/easy with mixed_qwen3vl SFT checkpoint.
+# VLM RL on VisGym (maze_2d/easy by default), starting from an SFT-trained
+# Qwen3-VL checkpoint that already emits the structured tuple-action format.
 #
 # Colocated setup: 8 GPUs shared between vLLM inference and FSDP training.
-# Uses the VLM generator (multi-modal) via custom entrypoint.
 #
-# Prerequisites – generate the stub dataset:
-#   uv run examples/train/visgym/visgym_dataset.py \
+# Prerequisites – generate the stub dataset (auto-created on first run):
+#   uv run examples/train/visgym/dataset.py \
 #       --env_id maze_2d/easy --num_rows 256 \
 #       --output_dir ~/data/visgym_maze_2d_easy
 #
 # Usage:
-#   bash examples/train/visgym/run_visgym_full_mazed2d_easy-finalized.sh
+#   MODEL_PATH=/path/to/sft_ckpt bash examples/train/visgym/run_visgym_from_sft.sh
 
 : "${ENV_ID:=maze_2d/easy}"
 : "${DATA_DIR:="$HOME/data/visgym_maze_2d_easy"}"
 : "${NUM_INFERENCE_GPUS:=8}"
 : "${NUM_TRAIN_GPUS:=8}"
 : "${LOGGER:=wandb}"
+: "${MODEL_PATH:=/mnt/local_storage/visgym_model_repo/mixed_qwen3vl}"
 
-: "${EXPORT_PATH:="$HOME/exports/visgym_maze_2d_easy_format_reward_0.2"}"
+: "${EXPORT_PATH:="$HOME/exports/visgym_maze_2d_easy_from_sft"}"
 : "${DUMP_TRAINING_BATCHES:=false}"
 : "${DUMP_EVAL_RESULTS:=true}"
 : "${EVAL_INTERVAL:=10}"
@@ -32,10 +33,9 @@ set -x
 : "${NUM_DATASET_ROWS:=256}"
 : "${EVAL_DATA_DIR:="$HOME/data/visgym_maze_2d_easy_eval"}"
 
-# ── Generate stub dataset if it doesn't exist ────────────────────────
 if [ ! -f "$DATA_DIR/train.parquet" ]; then
   echo "=== Generating train stub dataset for $ENV_ID ==="
-  uv run examples/train/visgym/visgym_dataset.py \
+  uv run examples/train/visgym/dataset.py \
     --env_id "$ENV_ID" \
     --num_rows "$NUM_DATASET_ROWS" \
     --output_dir "$DATA_DIR"
@@ -43,17 +43,16 @@ fi
 
 if [ ! -f "$EVAL_DATA_DIR/train.parquet" ]; then
   echo "=== Generating eval stub dataset for $ENV_ID ==="
-  uv run examples/train/visgym/visgym_dataset.py \
+  uv run examples/train/visgym/dataset.py \
     --env_id "$ENV_ID" \
     --num_rows 64 \
     --seed \
     --output_dir "$EVAL_DATA_DIR"
 fi
 
-MODEL_PATH="/mnt/local_storage/visgym_model_repo/mixed_qwen3vl"
 _SKYRL_USE_NEW_INFERENCE=1 \
 uv run --isolated --extra fsdp \
-  python examples/train/visgym/visgym_entrypoint.py \
+  python examples/train/visgym/entrypoint_sft.py \
   data.train_data="['$DATA_DIR/train.parquet']" \
   data.val_data="['$EVAL_DATA_DIR/train.parquet']" \
   trainer.algorithm.advantage_estimator="grpo" \
@@ -87,13 +86,13 @@ uv run --isolated --extra fsdp \
   environment.env_class=visgym \
   trainer.logger="$LOGGER" \
   trainer.project_name="vlm_maze_2d_easy" \
-  trainer.run_name="maze_2d_easy_full_format_reward_0.2-reduced_bsz" \
+  trainer.run_name="maze_2d_easy_from_sft" \
   trainer.resume_mode=null \
   trainer.log_path="/tmp/skyrl-logs" \
   trainer.export_path="$EXPORT_PATH" \
   trainer.dump_data_batch="$DUMP_TRAINING_BATCHES" \
   trainer.dump_eval_results="$DUMP_EVAL_RESULTS" \
-  trainer.ckpt_path="$HOME/ckpts/visgym_maze_2d_easy_format_reward_0.2" \
+  trainer.ckpt_path="$HOME/ckpts/visgym_maze_2d_easy_from_sft" \
   trainer.use_sample_packing=false \
   trainer.eval_interval="$EVAL_INTERVAL" \
   trainer.ckpt_interval=10 \
